Fix linting and import errors.
