+++ learn2learn/gym/envs/mujoco/humanoid_direction.py
@@ -7,7 +7,7 @@ from gym.error import DependencyNotInstalled
 try:
     from gym.envs.mujoco.mujoco_env import MujocoEnv
 except DependencyNotInstalled:
-    from learn2learn.gym.envs.mujoco import MujocoEnv
+    from learn2learn.gym.envs.mujoco.dummy_mujoco_env import MujocoEnv
 
 from learn2learn.gym.envs.meta_env import MetaEnv
